feat: раскидал функционвал игры на fsd
