+++ .prettierrc
@@ -3,5 +3,5 @@
   "semi": false,
   "arrowParens": "avoid",
   "bracketSameLine": true,
-  "width": 50
+  "width": 90
 }
