+++ packages/client/src/features/GameCanvas/lib/getRandomInt.ts
@@ -1,0 +1,3 @@
+export const getRandomInt = (min: number, max: number): number => {
+  return Math.floor(Math.random() * (max - min)) + min
+}
