+++ packages/client/src/features/GameCanvas/lib/index.ts
@@ -1,0 +1,1 @@
+export { getRandomInt } from './getRandomInt'
