+++ packages/client/src/features/GameCanvas/model/gameConstant.ts
@@ -1,0 +1,4 @@
+export const SIZE_OBJECT = 10
+export const COLOR_SNAKE = 'green'
+export const COLOR_APPLE = 'red'
+export const DEFAULT_SPEED = 200
