+++ packages/client/src/features/GameCanvas/model/hocs/index.ts
@@ -1,0 +1,2 @@
+export { useConfigurateCanvas } from './useCulcSizeCanvas'
+export { useGameControls } from './useGameControls'
