+++ packages/client/src/features/GameCanvas/model/types/ConfigCanvas.ts
@@ -1,0 +1,7 @@
+import { SizeCanvas } from '@/features/GameCanvas/model/types'
+
+export type ConfigCanvas = {
+  center: { x: number; y: number }
+  ctx: CanvasRenderingContext2D | null
+  sizeCanvas: SizeCanvas
+}
