+++ packages/client/src/features/GameCanvas/model/types/Direction.ts
@@ -1,0 +1,6 @@
+export enum Direction {
+  ArrowUp = 'ArrowUp',
+  ArrowDown = 'ArrowDown',
+  ArrowLeft = 'ArrowLeft',
+  ArrowRight = 'ArrowRight',
+}
