+++ packages/client/src/features/GameCanvas/model/types/GameItem.ts
@@ -1,0 +1,4 @@
+export type GameItem = {
+  x: number
+  y: number
+}
