+++ packages/client/src/features/GameCanvas/model/types/SizeCanvas.ts
@@ -1,0 +1,6 @@
+export type SizeCanvas = {
+  width: number
+  height: number
+  countCellWidth: number
+  countCellHeight: number
+}
