+++ packages/client/src/features/GameCanvas/model/types/Snake.ts
@@ -1,0 +1,3 @@
+import { GameItem } from '@/features/GameCanvas/model/types/GameItem'
+
+export type Snake = GameItem[]
