+++ packages/client/src/features/GameInfo/lib/index.ts
@@ -1,0 +1,1 @@
+export { formatTimeFromMS } from './formatedTimeFromMS'
