+++ packages/client/src/features/GameInfo/model/gameInfoTypes.ts
@@ -1,0 +1,3 @@
+export type GameInfoProps = {
+  scope: number
+}
