+++ packages/client/src/features/GameInfo/model/index.ts
@@ -1,0 +1,1 @@
+export type { GameInfoProps } from './gameInfoTypes'
