+++ packages/client/src/features/GameInfo/ui/index.ts
@@ -1,0 +1,1 @@
+export { GameInfo } from './GameInfo'
