--- packages/client/src/pages/GamePage/GamePage.module.css
@@ -1,5 +1,0 @@
-.canvas-game {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
