+++ packages/client/src/widgets/index.ts
@@ -1,1 +1,2 @@
 export { MainLayout } from './layouts/MainLayout'
+export { Game } from './Game/Game'
